fix(ui): address 5 design issues from post-redesign review
Fix 1+2: OPS panel MCP calls + non-blocking UI
- Add sirin_mcp_call() to IntegrationService trait + impl: calls
  http://127.0.0.1:7700/mcp directly (Sirin's own tools, not external MCPs)
- ops_panel.rs rewritten with async pattern: spawn_sirin_call() runs
  MCP in background thread, writes into Arc<Mutex<Option<Result>>>
  ResultCell, requests repaint when done → no UI freeze on LLM calls
- result_box() helper: shows Spinner while running, result on done
- All buttons disabled while their call is in-flight

Fix 3: Icon consistency in icon-only sidebar
- Remove emoji (⚡📋⚙) — unstable size/alignment in egui
- Use existing drawn icons (Gear/Monitor/Globe/People) for new entries
- OPS icons use single ASCII Glyph("A"/"T"/"$") — consistent with agent letters

Fix 4: ensure_view_visible for new View variants
- Coverage/BrowserMonitor → testing_open = true
- Team/McpPlayground → automation_open = true
- AiRouter/SessionTasks/CostKb → ops_open = true
- Clicking any view in icon mode auto-expands correct section in labelled mode

Fix 5: coverage_color visibility
- Changed pub fn → fn (module-private) to avoid cross-module name confusion

+++ src/browser_ax.rs
@@ -287,7 +287,7 @@ fn collect_hidden_backend_ids() -> Result<HashSet<u32>, String> {
                     depth: Some(0),
                     pierce: Some(false),
                 }) {
-                    let bid = desc.node.backend_node_id as u32;
+                    let bid = desc.node.backend_node_id;
                     out.insert(bid);
                 }
             }
